+++ taxi/commands/base.py
@@ -60,9 +60,11 @@ def populate_backends(backends, context):
     plugins_registry.populate_backends(dict(backends), context)
 
 
-def create_config_file(filename):
+def create_config_file(filename, run_conversions=True):
     """
-    Create main configuration file if it doesn't exist.
+    Create main configuration file if it doesn't exist. If `run_conversions` is
+    set and the configuration file already exists, it is converted to the latest
+    config file format.
     """
     import textwrap
     from urllib import parse
@@ -141,7 +143,7 @@ def create_config_file(filename):
 
         with open(filename, 'w') as f:
             f.write(templated_config)
-    else:
+    elif run_conversions:
         settings = Settings(filename)
         conversions = settings.needed_conversions
 
@@ -285,8 +287,9 @@ def cli(ctx, config, taxi_dir, verbose):
     logging.debug("Using configuration file in %s", config)
     logging.debug("Using data directory %s", taxi_dir)
 
-    create_config_file(config)
-    settings = Settings(config)
+    is_config = ctx.invoked_subcommand == "config"
+    create_config_file(config, run_conversions=not is_config)
+    settings = Settings(config) if not is_config else None
     projects_db = ProjectsDb(os.path.expanduser(taxi_dir))
 
     if not os.path.exists(taxi_dir):
@@ -296,9 +299,11 @@ def cli(ctx, config, taxi_dir, verbose):
     ctx.obj['settings'] = settings
     ctx.obj['view'] = TtyUi()
     ctx.obj['projects_db'] = projects_db
+    ctx.obj['config_path'] = config
 
-    populate_aliases(get_all_aliases(projects_db=projects_db, settings=settings))
-    populate_backends(settings.get_backends(), ctx.obj)
+    if not is_config:
+        populate_aliases(get_all_aliases(projects_db=projects_db, settings=settings))
+        populate_backends(settings.get_backends(), ctx.obj)
 
 
 # This can't be called from inside a command because Click will already have built its commands list